+++ core/src/main/java/io/questdb/cairo/MicrosTimestampDriver.java
@@ -390,7 +390,7 @@ public class MicrosTimestampDriver implements TimestampDriver {
     }
 
     @Override
-    public long fromSeconds(int seconds) {
+    public long fromSeconds(long seconds) {
         return seconds * Timestamps.SECOND_MICROS;
     }
